Remove secrets from the Handel-Codepipeline config
This way you can commit this file to your repository in case you
ever need to create a new pipeline from it.

The initial prompts now ask you for your GitHub access token and
the account ID you want to deploy the pipeline to, as well as
the directory containing the account configurations. The secrets
are not persisted anywhere.