Allow CodeBuild to access EC2 Parameter Store for app's params
This change grants the build phase of a pipeline access to the
parameters in EC2 Parameter Store under the app's namespace. If
the Handel app name is 'myapp', then it will have access to any
parameter of the format 'myapp*'.

This change also fixes the bug where codebuild and codepipeline
creations would fail because newly created roles werent available
yet.