Remove Handel dependency from Handel-CodePipeline
Previously, we required you to already have a Handel file created.
The only piece of information we got from this was the 'name' field.

This change removes that requirement to have a Handel file, and
instead adds a 'name' field to the handel-codepipeline.yml file.

This will allow people to use Handel-CodePipeline to create pipelines
that don't even use Handel, such as a pipeline that just has source
and build phases.