Add initial support for Handel Delete phase
This phase deletes an environment that was previously created in
a pipeline. This is intended for use in ephemeral environments that
only need to persist during something like tests in a pipeline.

There's a giant warning in the docs for this phase that state how
dangerous this phase is, because it deletes all data.