Fix bug where double-slash would get appended to image name.
The documentation tells users to include the slash between repo
location and repo name, but the code was also doing this. We remove
the code doing this, instead requiring the user to put it in their
image name. It looks better in their handel-codepipeline.yml file
to have the slash included like "<account>/myrepo".